Dynamic blog from scratch
Finalized dynamic blog:
Using EJS, express and node.
All pages imports header and footer from partials.
Home page has a static hero image and text.
Home page imports the 3 newest posts from the blog array to populate it.
The posts page does the same as the home page, however it'll render all the posts on the posts array.
The admin page checks if the user is logged in, if true renders the admin page, where the user will be able to create posts. If false renders the login page asking for username and password (admin, 123 hard coded)